Updated to 1.15.1

+++ src/main/kotlin/com/scientianovateam/versatile/machines/gui/textures/interactable/ButtonTextureGroup.kt
@@ -3,14 +3,14 @@ package com.scientianovateam.versatile.machines.gui.textures.interactable
 import com.scientianovateam.versatile.machines.gui.textures.IDrawable
 
 data class ButtonTextureGroup @JvmOverloads constructor(val on: IDrawable, val off: IDrawable, val selected: IDrawable = off) : IDrawable {
-    override val uStart = 0.0
-    override val vStart = 0.0
-    override val uEnd = 1.0
-    override val vEnd = 1.0
+    override val uStart = 0f
+    override val vStart = 0f
+    override val uEnd = 1f
+    override val vEnd = 1f
 
-    override fun draw(x: Int, y: Int, width: Int, height: Int, uStart: Double, vStart: Double, uEnd: Double, vEnd: Double) = draw(x, y, width, height, uStart, vStart, uEnd, vEnd, ButtonDrawMode.OFF)
+    override fun draw(x: Int, y: Int, width: Int, height: Int, uStart: Float, vStart: Float, uEnd: Float, vEnd: Float) = draw(x, y, width, height, uStart, vStart, uEnd, vEnd, ButtonDrawMode.OFF)
 
-    fun draw(x: Int, y: Int, width: Int, height: Int, uStart: Double = 0.0, vStart: Double = 0.0, uEnd: Double = 1.0, vEnd: Double = 1.0, drawMode: ButtonDrawMode) = (when (drawMode) {
+    fun draw(x: Int, y: Int, width: Int, height: Int, uStart: Float = 0f, vStart: Float = 0f, uEnd: Float = 1f, vEnd: Float = 1f, drawMode: ButtonDrawMode) = (when (drawMode) {
         ButtonDrawMode.OFF -> on
         ButtonDrawMode.SELECTED -> selected
         ButtonDrawMode.ON -> off
